+++ hydra-node/test/Hydra/Ledger/CardanoSpec.hs
@@ -14,6 +14,7 @@ import Data.Aeson (eitherDecode, encode)
 import qualified Data.Aeson as Aeson
 import qualified Data.ByteString.Base16 as Base16
 import Data.Text (unpack)
+import Hydra.Cardano.Api.Pretty (renderTx)
 import Hydra.Chain.Direct.Fixture (defaultGlobals, defaultLedgerEnv)
 import Hydra.Ledger (applyTransactions)
 import Hydra.Ledger.Cardano (
@@ -23,7 +24,6 @@ import Hydra.Ledger.Cardano (
   genUTxOAdaOnlyOfSize,
   genUTxOAlonzo,
   genUTxOFor,
-  renderTx,
  )
 import Hydra.Ledger.Cardano.Evaluate (slotNoFromUTCTime, slotNoToUTCTime)
 import Test.Aeson.GenericSpecs (roundtripAndGoldenSpecs)